+++ app/src/test/kotlin/fr/free/nrw/commons/profile/achievements/AchievementsFragmentUnitTests.kt
@@ -1,9 +1,7 @@
 package fr.free.nrw.commons.profile.achievements
 
 import android.content.Context
-import android.graphics.Bitmap
 import android.os.Looper
-import android.view.MenuItem
 import android.widget.ImageView
 import android.widget.ProgressBar
 import android.widget.TextView
@@ -26,7 +24,6 @@ import org.robolectric.RuntimeEnvironment
 import org.robolectric.Shadows
 import org.robolectric.annotation.Config
 import org.robolectric.annotation.LooperMode
-import org.robolectric.fakes.RoboMenuItem
 import org.wikipedia.AppAdapter
 import java.lang.reflect.Method
 
@@ -39,8 +36,6 @@ class AchievementsFragmentUnitTests {
     private lateinit var fragment: AchievementsFragment
 
     private lateinit var context: Context
-
-    private lateinit var menuItem: MenuItem
 
     private lateinit var achievements: Achievements
 
@@ -84,16 +79,12 @@ class AchievementsFragmentUnitTests {
     private lateinit var imageUploadedText: TextView
 
     @Mock
-    private lateinit var bitmap: Bitmap
-
-    @Mock
     private lateinit var progressBar: ProgressBar
 
     @Before
     fun setUp() {
         MockitoAnnotations.initMocks(this)
         context = RuntimeEnvironment.application.applicationContext
-        menuItem = RoboMenuItem(context)
         AppAdapter.set(TestAppAdapter())
         val activity = Robolectric.buildActivity(ProfileActivity::class.java).create().get()
         fragment = AchievementsFragment()
@@ -134,7 +125,6 @@ class AchievementsFragmentUnitTests {
         Whitebox.setInternalState(fragment, "imageView", imageView)
         Whitebox.setInternalState(fragment, "progressBar", progressBar)
         Whitebox.setInternalState(fragment, "imagesRevertLimitText", imagesRevertLimitText)
-        Whitebox.setInternalState(fragment, "item", menuItem)
     }
 
     @Test
@@ -142,14 +132,6 @@ class AchievementsFragmentUnitTests {
     fun checkFragmentNotNull() {
         Assert.assertNotNull(fragment)
     }
-
-    @Test
-    @Throws(Exception::class)
-    fun testShowAlert() {
-        Shadows.shadowOf(Looper.getMainLooper()).idle()
-        fragment.showAlert(bitmap)
-    }
-
 
     @Test
     @Throws(Exception::class)
@@ -205,13 +187,6 @@ class AchievementsFragmentUnitTests {
     fun testShowQualityImagesInfo() {
         Shadows.shadowOf(Looper.getMainLooper()).idle()
         fragment.showQualityImagesInfo()
-    }
-
-    @Test
-    @Throws(Exception::class)
-    fun testOnOptionsItemSelected() {
-        Shadows.shadowOf(Looper.getMainLooper()).idle()
-        fragment.onOptionsItemSelected(menuItem)
     }
 
     @Test
@@ -329,16 +304,4 @@ class AchievementsFragmentUnitTests {
         method.isAccessible = true
         method.invoke(fragment)
     }
-
-    @Test
-    @Throws(Exception::class)
-    fun testShareScreen() {
-        Shadows.shadowOf(Looper.getMainLooper()).idle()
-        val method: Method = AchievementsFragment::class.java.getDeclaredMethod(
-            "shareScreen", Bitmap::class.java
-        )
-        method.isAccessible = true
-        method.invoke(fragment, bitmap)
-    }
-
 }
